Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -3,7 +3,7 @@ class CreateCostumes < ActiveRecord::Migration[5.2]
   def change
     create_table :costumes do |c|
       c.string :name
-      c.float :price
+      c.integer :price
       c.string :image_url
       c.string :size
       c.datetime :created_at
